requirements.txt
Django==4.1.7
environs==9.5.*
Pillow==9.3.0
python-dateutil==2.8.2
django-multiselectfield==0.1.12
gunicorn==20.0.2
django-widget-tweaks==1.4.12
psycopg2-binary==2.9.5
requests==2.28.*